requirements.txt
Django==2.2.4
django-filter==2.2.0
pytz==2019.2
sqlparse==0.3.0
beautifulsoup4==4.8.0
Django==2.2.4
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==1.0.1
django-crispy-forms==1.7.2
django-filter==2.2.0
pytz==2019.2
soupsieve==1.9.3
sqlparse==0.3.0
